requirements.txt
# Core Framework
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Google Gemini API
google-generativeai>=0.3.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
reportlab>=4.0.0

# Text Processing & Embeddings
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# Vector Database
chromadb>=0.4.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
tqdm>=4.66.0
loguru>=0.7.0
python-multipart>=0.0.6

# Additional dependencies for stability
Pillow>=10.0.0
requests>=2.31.0